Move utils, widgets and constants folders to src
