+++ lib/src/utils/text_utils.dart
