+++ lib/src/widgets/checkbox_group.dart
@@ -1,0 +1,112 @@
+/// Checkbox group widget.
+///
+// Time-stamp: <Sunday 2023-12-31 18:58:28 +1100 Graham Williams>
+///
+/// Copyright (C) 2024, Software Innovation Institute, ANU.
+///
+/// Licensed under the MIT License (the "License").
+///
+/// License: https://choosealicense.com/licenses/mit/.
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+///
+/// Authors: Tony Chen
+
+import 'package:flutter/material.dart';
+import 'package:markdown_widgets/src/constants/pkg.dart'
+    show contentWidthFactor;
+
+class CheckboxGroup extends StatefulWidget {
+  final String name;
+  final List<Map<String, String>> options;
+  final Set<String> selectedValues;
+  final ValueChanged<Set<String>> onChanged;
+
+  const CheckboxGroup({
+    Key? key,
+    required this.name,
+    required this.options,
+    required this.selectedValues,
+    required this.onChanged,
+  }) : super(key: key);
+
+  @override
+  _CheckboxGroupState createState() => _CheckboxGroupState();
+}
+
+class _CheckboxGroupState extends State<CheckboxGroup> {
+  late Set<String> _selectedValues;
+
+  @override
+  void initState() {
+    super.initState();
+    _selectedValues = Set<String>.from(widget.selectedValues);
+  }
+
+  @override
+  void didUpdateWidget(CheckboxGroup oldWidget) {
+    super.didUpdateWidget(oldWidget);
+    if (widget.selectedValues != oldWidget.selectedValues) {
+      setState(() {
+        _selectedValues = Set<String>.from(widget.selectedValues);
+      });
+    }
+  }
+
+  void _onChanged(String value, bool? isChecked) {
+    setState(() {
+      if (isChecked == true) {
+        _selectedValues.add(value);
+      } else {
+        _selectedValues.remove(value);
+      }
+    });
+    widget.onChanged(_selectedValues);
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Center(
+      child: FractionallySizedBox(
+        widthFactor: contentWidthFactor,
+        child: Column(
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: widget.options.map((option) {
+            bool isChecked = _selectedValues.contains(option['value']!);
+            return Row(
+              children: [
+                Checkbox(
+                  value: isChecked,
+                  onChanged: (bool? newValue) {
+                    _onChanged(option['value']!, newValue);
+                  },
+                ),
+                Expanded(
+                  child: Text(
+                    option['label']!,
+                  ),
+                ),
+              ],
+            );
+          }).toList(),
+        ),
+      ),
+    );
+  }
+}
