+++ lib/src/widgets/radio_group.dart
@@ -1,4 +1,4 @@
-/// Slider bar widget.
+/// Radio button group widget.
 ///
 // Time-stamp: <Sunday 2023-12-31 18:58:28 +1100 Graham Williams>
 ///
@@ -29,24 +29,20 @@
 /// Authors: Tony Chen
 
 import 'package:flutter/material.dart';
-import 'package:markdown_widgets/constants/pkg.dart'
+import 'package:markdown_widgets/src/constants/pkg.dart'
     show contentWidthFactor;
 
-class SliderWidget extends StatelessWidget {
+class RadioGroup extends StatelessWidget {
   final String name;
-  final double value;
-  final double min;
-  final double max;
-  final double step;
-  final ValueChanged<double> onChanged;
+  final List<Map<String, String>> options;
+  final String? selectedValue;
+  final ValueChanged<String?> onChanged;
 
-  const SliderWidget({
+  const RadioGroup({
     Key? key,
     required this.name,
-    required this.value,
-    required this.min,
-    required this.max,
-    required this.step,
+    required this.options,
+    this.selectedValue,
     required this.onChanged,
   }) : super(key: key);
 
@@ -55,13 +51,24 @@ class SliderWidget extends StatelessWidget {
     return Center(
       child: FractionallySizedBox(
         widthFactor: contentWidthFactor,
-        child: Slider(
-          value: value,
-          min: min,
-          max: max,
-          divisions: ((max - min) / step).round(),
-          label: value.toStringAsFixed(0),
-          onChanged: onChanged,
+        child: Column(
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: options.map((option) {
+            return Row(
+              children: [
+                Radio<String>(
+                  value: option['value']!,
+                  groupValue: selectedValue,
+                  onChanged: onChanged,
+                ),
+                Expanded(
+                  child: Text(
+                    option['label']!,
+                  ),
+                ),
+              ],
+            );
+          }).toList(),
         ),
       ),
     );
