+++ lib/src/widgets/text_heading.dart
@@ -29,8 +29,8 @@
 /// Authors: Tony Chen
 
 import 'package:flutter/material.dart';
-import 'package:markdown_widgets/utils/text_utils.dart';
-import 'package:markdown_widgets/constants/pkg.dart'
+import 'package:markdown_widgets/src/utils/text_utils.dart';
+import 'package:markdown_widgets/src/constants/pkg.dart'
     show contentWidthFactor, screenWidth;
 
 class TextHeadingWidget extends StatelessWidget {
